ingestion-phase/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
torch>=2.0.0
chromadb
endee>=0.1.8
requests>=2.28.0
pymongo>=4.0.0
transformers>=4.30.0
pyautogen>=0.2.0
numpy>=1.24.0
tqdm>=4.65.0
dnspython>=2.3.0
bert-score>=0.3.13
huggingface-hub>=0.16.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# PDF generation and template overlay
reportlab>=4.0.0
pypdf>=4.2.0

# Editable document export
python-docx>=0.8.11

# FastAPI and async support
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
motor>=3.3.0

# Groq API client
groq>=0.4.0
python-dotenv>=1.0.0

